Fix casting on MSU1 volume.

+++ msu1.cpp
@@ -433,8 +433,9 @@ void S9xMSU1Generate(size_t sample_count)
 			int bytes_read = READ_STREAM((char *)&sample, 4, audioStream);
 			if (bytes_read == 4)
 			{
-				*left =  (int32) GET_LE16(left)  * MSU1.MSU1_VOLUME / 255;
-				*right = (int32) GET_LE16(right) * MSU1.MSU1_VOLUME / 255;
+				*left = ((int32)(int16)GET_LE16(left) * MSU1.MSU1_VOLUME / 255);
+				*right = ((int32)(int16)GET_LE16(right) * MSU1.MSU1_VOLUME / 255);
+
 
 				*(bufPos++) = *left;
 				*(bufPos++) = *right;
